+++ guillotina_gcloudstorage/storage.py
@@ -179,8 +179,13 @@ class GCloudFileManager(object):
         return resp
 
     async def tus_patch(self):
-        self.context._p_register()  # writing to object
         file = self.field.get(self.field.context or self.context)
+
+        try:
+            self.field.context.data._p_register()  # register change...
+        except AttributeError:
+            self.context._p_register()
+
         if 'CONTENT-LENGTH' in self.request.headers:
             to_upload = int(self.request.headers['CONTENT-LENGTH'])
         else: